% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{BIC.brmsfit}
\alias{BIC.brmsfit}
\alias{brmsfit}
\title{Bayesian Information Criterion for \code{stanreg} model object.}
\usage{
\method{BIC}{brmsfit}(object, estimate = c("Mean", "MAP"), ...)
}
\arguments{
\item{object}{an object returned by \code{brm()} function.}

\item{estimate}{a string specifying the estimate to use in evaluating the
log-likelihood: \code{"Mean"} for the posterior mean, \code{"MAP"} for the maximum
a posteriori.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
The value of BIC.
}
\description{
Return BIC for fitted \code{brmsfit} model.
}
\details{
Only available for \code{family = gaussian(link = "identity")}.
}
\examples{
x = rnorm(100)
y = 1 + 0.5*x + rnorm(100, 0, 0.5)
mod = brm(y ~ x, data = data.frame(x, y), 
          family = gaussian(link = "identity"))
BIC(mod)

}
