% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{compareModels}
\alias{compareModels}
\title{Models comparison based on information criteria for \code{stanreg}
model objects.}
\usage{
compareModels(..., criterion = c("BIC", "DIC", "WAIC"))
}
\arguments{
\item{\dots}{model objects returned by \code{stan_glm()} function.}

\item{criterion}{a string specifying the criterion to use.}
}
\value{
Return a \code{data.table} of selected information criterion,
the effective number of estimated parameters, and the criterion
difference with the one with the smallest value.
}
\description{
Compare fitted models using a selected information criterion.
}
\details{
Only available for \code{family = gaussian(link = "identity")}.
}
\examples{
x = rnorm(100)
y = 1 + 0.5*x + rnorm(100, 0, 0.5)

\dontrun{
mod1 = stan_glm(y ~ 1, data = data.frame(x, y))
mod2 = stan_glm(y ~ x, data = data.frame(x, y))
mod3 = stan_glm(y ~ x + I(x^2), data = data.frame(x, y))
compareModels(mod1, mod2, mod3, criterion = "BIC")
compareModels(mod1, mod2, mod3, criterion = "DIC")
compareModels(mod1, mod2, mod3, criterion = "WAIC")
}

Mod1 = brm(y ~ 1, data = data.frame(x, y))
Mod2 = brm(y ~ x, data = data.frame(x, y))
Mod3 = brm(y ~ x + I(x^2), data = data.frame(x, y))
compareModels(Mod1, Mod2, Mod3, criterion = "BIC")
compareModels(Mod1, Mod2, Mod3, criterion = "DIC")
compareModels(Mod1, Mod2, Mod3, criterion = "WAIC")

}
