% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Summary statistics for numerical data}
\usage{
describe(x, probs = c(0, 0.25, 0.75, 1), ...)
}
\arguments{
\item{x}{a vector (univariate), data.frame or matrix (multivariate case)
of numerical values.}

\item{probs}{a numeric vector of probabilities with values in [0,1] used
for computing quantiles.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A `data.frame` containing a description of the dataset.
}
\description{
Compute summary statistics for the input data.
}
\examples{

x = rnorm(200, 10, 5)
describe(x)
xy = data.frame(x = rnorm(200, 10, 5), y = rchisq(200, 3))
describe(xy)
describe(xy[,"y"])

}
