% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{DIC.brmsfit}
\alias{DIC.brmsfit}
\title{Deviance Information Criterion}
\usage{
\method{DIC}{brmsfit}(object, estimate = c("Mean", "MAP"), ...)
}
\arguments{
\item{object}{an object returned by \code{brm()} function.}

\item{estimate}{a string specifying the point estimate to use.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
The value of DIC.
}
\description{
Return DIC for fitted \code{brmsfit} model.
}
\details{
Only available for \code{family = gaussian(link = "identity")}.
}
\examples{
x = rnorm(100)
y = 1 + 0.5*x + rnorm(100, 0, 0.5)
mod = brm(y ~ x, data = data.frame(x, y), 
          family = gaussian(link = "identity"))
DIC(mod)

}
