% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{\dots}{other arguments.}
}
\value{
The value of DIC.
}
\description{
Generic method for computing Deviance Information
Criterion (DIC)
}
