% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{DIC.stanreg}
\alias{DIC.stanreg}
\title{Deviance Information Criterion}
\usage{
\method{DIC}{stanreg}(object, estimate = c("Mean", "MAP"), ...)
}
\arguments{
\item{object}{an object returned by \code{stan_glm()} function.}

\item{estimate}{a string specifying the point estimate to use.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
The value of DIC.
}
\description{
Return DIC for fitted \code{stanreg} model.
}
\details{
Only available for \code{family = gaussian(link = "identity")}.
}
\examples{
\dontrun{
x = rnorm(100)
y = 1 + 0.5*x + rnorm(100, 0, 0.5)
mod = stan_glm(y ~ x, data = data.frame(x, y))
DIC(mod)
}

}
