% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mcmc.R
\name{HPD_interval}
\alias{HPD_interval}
\title{Highest posterior density (HPD) intervals from MCMC sampling}
\usage{
HPD_interval(draws, level = 0.95, ...)
}
\arguments{
\item{draws}{a vector of MCMC draws}

\item{level}{a numerical value in (0,1) specifying the nominal level of credible interval}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A named list containing
\itemize{
\item \code{lower} the lower limit of credible interval
\item \code{upper} the upper limit of credible interval
\item \code{level} the nominal level of credible interval
\item \code{probr} the empirical probability level of computed credible interval
}
}
\description{
Computes the highest posterior density intervals from
MCMC sample draws at specified probability level.
}
\examples{
x = rnorm(1e7, 0, 1)
HPD_interval(x)

x = rchisq(1e7, df = 5)
HPD_interval(x)

}
