% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mcmc.R
\name{plotDiagnostic}
\alias{plotDiagnostic}
\title{Diagnostic plots for MCMC sampling}
\usage{
plotDiagnostic(
  object,
  pars = NULL,
  what = c("trace", "dens", "acf"),
  lag = 20,
  ...
)
}
\arguments{
\item{object}{an object of class \code{CmdStanMCMC} or \code{brmsfit}.}

\item{pars}{an optional character vector of parameter names to monitor.}

\item{what}{a string specifying the type of diagnostic plot to draw; available options are \code{"trace"}, \code{"dens"}, and \code{"acf"}.}

\item{lag}{a numerical value for the maximum lag to use when computing autocorrelation in ACF plot.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A \code{ggplot2} object containing the requested diagnostic plot.
}
\description{
Draw diagnostic plots to be used for checking MCMC convergence.
}
\details{
The resulting plots are inspired by those used in
Kruschke J.K. (2015) \emph{Doing Bayesian Data Analysis}, 2nd ed.
}
\examples{
data = list(y = 352, n = 1012, alpha = 2, beta = 2)
stan_file = file.path(system.file(package = "bayadvstat"), "stan", "beta-binomial.stan")
model = cmdstan_model(stan_file)
fit = model$sample(data = data)
plotDiagnostic(fit, what = "trace")
plotDiagnostic(fit, what = "dens")
plotDiagnostic(fit, what = "acf")
# use patchwork to collect graphs:
plotDiagnostic(fit, what = "trace") +
  plotDiagnostic(fit, what = "dens") +
  plotDiagnostic(fit, what = "acf") +
  plot_layout(guides = "collect") 
  
stan_file = file.path(system.file(package = "bayadvstat"), "stan", "normal-normal.stan")
model = cmdstan_model(stan_file)
y = c(9.37, 10.18, 9.16, 11.60, 10.33)
data = list(y = y, n = length(y),                  # data
            mu_prior = 5, sigma_prior = sqrt(10),  # prior hyperparameters on mu
            tau = 1)                               # prior hyperparameters on sigma
fit = model$sample(data = data,
                   iter_warmup = 1000,
                   iter_sampling = 4000,
                   chains = 4)
plotDiagnostic(fit, pars = "mu", what = "trace")
plotDiagnostic(fit, pars = "sigma", what = "dens")
plotDiagnostic(fit, pars = "mu", what = "acf")
wrap_plots(plotDiagnostic(fit, what = "trace"), ncol = 1)

x = rnorm(100)
y = 1 + 0.5 * x + rnorm(100)
mod = brm(y ~ x, data = data.frame(x, y), 
          family = gaussian())
p = plotDiagnostic(mod, pars = c("b_Intercept", "b_x", "sigma"), 
                   what = "trace")
wrap_plots(p) + plot_layout(guides = "collect") 
   
}
