% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mcmc.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Posterior plots for MCMC sampling}
\usage{
plotPosterior(
  object,
  pars = NULL,
  center = c("mean", "map", "median"),
  level = 0.95,
  digits = min(4, getOption("digits")),
  ...
)
}
\arguments{
\item{object}{an object of class \code{CmdStanMCMC} or \code{brmsfit}.}

\item{pars}{an optional character vector of parameter names to monitor.}

\item{center}{a string specifying the type of center statistics to draw; available options are \code{"mean"}, \code{"map"}, and \code{"median"}.}

\item{level}{a numerical value in (0,1) specifying the level of credible intervals.}

\item{digits}{significant digits to use for showing posterior statistics.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A \code{ggplot2} object (or a list of object) containing the requested
posterior plot(s).
}
\description{
Draw posterior plots for parameters from MCMC convergence.
}
\details{
The resulting plot is inspired by those used in
Kruschke J.K. (2015) \emph{Doing Bayesian Data Analysis}, 2nd ed.
}
\examples{
stan_file = file.path(system.file(package = "bayadvstat"), "stan", "beta-binomial.stan")
model = cmdstan_model(stan_file)
data = list(y = 352, n = 1012, alpha = 2, beta = 2)
fit = model$sample(data = data)
plotPosterior(fit)
plotPosterior(fit, center = "map")
plotPosterior(fit, level = 0.9)

stan_file = file.path(system.file(package = "bayadvstat"), "stan", "normal-normal.stan")
model = cmdstan_model(stan_file)
y = c(9.37, 10.18, 9.16, 11.60, 10.33)
data = list(y = y, n = length(y),                  # data
            mu_prior = 5, sigma_prior = sqrt(10),  # prior hyperparameters on mu
            tau = 1)                               # prior hyperparameters on sigma
fit = model$sample(data = data,
                   iter_warmup = 1000,
                   iter_sampling = 4000,
                   chains = 4)
plotPosterior(fit, pars = "mu")
plotPosterior(fit, pars = "sigma")
patchwork::wrap_plots(plotPosterior(fit))

x = rnorm(100)
y = 1 + 0.5 * x + rnorm(100)
mod = brm(y ~ x, data = data.frame(x, y), 
          family = gaussian())
plotPosterior(mod, pars = c("b_Intercept", "b_x", "sigma"))

}
