% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mcmc.R
\name{sample_mode}
\alias{sample_mode}
\title{Sample mode}
\usage{
sample_mode(x, ...)
}
\arguments{
\item{x}{a vector of data values}

\item{\dots}{additional arguments to be passed to base \code{\link[=density]{density()}} function for computing KDE.}
}
\value{
A numerical value.
}
\description{
Computes the sample mode.
}
\examples{
x = rnorm(1000, 10, 1)
sample_mode(x)

}
