% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mcmc.R
\name{summary.CmdStanFit}
\alias{summary.CmdStanFit}
\alias{print.summary.CmdStanFit}
\title{Summary of MCMC parameters sampling}
\usage{
\method{summary}{CmdStanFit}(object, pars = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{'CmdStanFit'} obtained by sampling
from \code{CmdStanModel} via \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{pars}{Character vector of parameters to include.}

\item{level}{The level of the credible intervals.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns a \code{data.frame}.
}
\description{
Summary statistics for parameters from MCMC sampling via Stan.
}
\examples{
data = list(y = 352, n = 1012, alpha = 2, beta = 2)
stan_file = file.path(system.file(package = "bayadvstat"), "stan", "beta-binomial.stan")
model = cmdstan_model(stan_file)
fit = model$sample(data = data)
summary(fit)

stan_file = file.path(system.file(package = "bayadvstat"), "stan", "normal-normal.stan")
model = cmdstan_model(stan_file)
y = c(9.37, 10.18, 9.16, 11.60, 10.33)
data = list(y = y, n = length(y),                  # data
            mu_prior = 5, sigma_prior = sqrt(10),  # prior hyperparameters on mu
            tau = 1)                               # prior hyperparameters on sigma
fit = model$sample(data = data,
                   iter_warmup = 1000,
                   iter_sampling = 4000,
                   chains = 4)
summary(fit)
                    
}
