% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{WAIC.stanreg}
\alias{WAIC.stanreg}
\title{Watanabe-Akaike Information Criterion for \code{stanreg} model object.}
\usage{
\method{WAIC}{stanreg}(object, ...)
}
\arguments{
\item{object}{an object returned by \code{stan_glm()} function.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
The value of WAIC.
}
\description{
Return WAIC for fitted \code{stanreg} model.
}
\details{
Only available for \code{family = gaussian(link = "identity")}.
}
\examples{
\dontrun{
x = rnorm(100)
y = 1 + 0.5*x + rnorm(100, 0, 0.5)
mod = stan_glm(y ~ x, data = data.frame(x, y))
WAIC(mod)
}

}
