% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{print.describe}
\title{Descriptive statistics}
\usage{
describe(data, by, probs = c(0, 0.25, 0.5, 0.75, 1), ...)

\method{print}{describe}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{data}{a matrix, data.frame, or vector.}

\item{by}{the name of a variable to condition on,}

\item{probs}{a vector of values in \eqn{[0,1]} specifying the quantiles to
be computed.}

\item{\dots}{additional arguments.}

\item{x}{an object of class \code{describe}.}

\item{digits}{significant digits.}
}
\value{
A \code{data.frame} containing a description of the dataset.
}
\description{
Compute descriptive statistics for input data.
}
\examples{

describe(iris)
describe(iris, by = Species)

}
