% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsel.R
\name{WAIC}
\alias{WAIC}
\title{Watanabe-Akaike Information Criterion}
\usage{
WAIC(object, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{\dots}{other arguments.}
}
\value{
The value of WAIC.
}
\description{
Generic method for computing Watanabe-Akaike
Information Criterion (WAIC)
}
